toudou_cli_starter/pyproject.toml
[project]
name = "Toudou"
version = "0.1"
description = "A simple task manager."
readme = "README.md"
dependencies = [
    "click>=8.1.7",
    "sqlalchemy>=2.0.27",
    "Flask>=3.0.2",
    "python-dotenv>=1.0.1",
    "flask-wtf>=1.2.1",
    "WTForms>=3.1.2",
    "flask-httpauth>=4.8.0",
]
requires-python = ">=3.10"
license = { text = "MIT" }


[project.scripts]
toudou = "toudou.views:cli"


[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.scripts]
start = "Flask --app toudou.vue.viewsWeb --debug run"

[tool.pdm.scripts.env]
path = "dev.env"
